Don't hold on to pygit2 Blob objects, and make cache optional
The reason we don't want to hold on to pygit2 Blob objects is that that
prevents the underlying git_blob_free() from being called right after we
are doing counting lines for a blob.

This is bad, since memoryview(blob) calls the libgi2
git_blob_rawcontent(), which very quickly fills up hunder of megs of
heap for large repos.

With this refactoring, we get this as the program runs:

    git_blob_rawcontent()
    git_object_free() // equivalent to git_blob_free()
    git_blob_rawcontent()
    git_object_free()
    git_blob_rawcontent()
    git_object_free()

instead of this

    git_blob_rawcontent()
    git_blob_rawcontent()
    git_blob_rawcontent()
    git_object_free()
    git_object_free()
    git_object_free()